+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file.
 
+### [Unreleased][unreleased]
+
+### [v0.0.2] - 2017-04-12
+
+#### Improvement
+
+- Remove property "defaultValue" in ActionInput.
+- Add Documentation.
+
+#### Fix
+
+- Debug when create interactor with combo value empty.
+
 ### [v0.0.1] - 2016-12-01
 
 - Create unique addAction function to add action.
@@ -14,6 +27,7 @@ All notable changes to this project will be documented in this file.
 - Recover an interactor
 - Interactor and action validation
 
-[unreleased]: https://github.com/AdrienEtienne/actions-interactor/compare/v0.0.1...HEAD
+[unreleased]: https://github.com/AdrienEtienne/actions-interactor/compare/v0.0.2...HEAD
+[v0.0.2]: https://github.com/AdrienEtienne/actions-interactor/compare/v0.0.1...v0.0.2
 [v0.0.1]: https://github.com/AdrienEtienne/actions-interactor/compare/v0.0.0...v0.0.1
 [v0.0.0]: https://github.com/AdrienEtienne/actions-interactor/compare/5bdd04c...v0.0.0